Ensure notification repository is discovered

+++ api/src/main/java/com/example/notification/repository/NotificationMasterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface NotificationMasterRepository extends JpaRepository<NotificationMaster, Long> {
+public interface NotificationMasterRepository extends JpaRepository<NotificationMaster, Integer> {
     Optional<NotificationMaster> findByCodeAndIsActiveTrue(String code);
 }
